Flask/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
